+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Adds some new methods to the env ("_") object given to a node's JavaScript snipp
  * _.addOption(label, target, value) adds a new option to the options menu of the node
  * Using next or back on nodes with options now works, too
  * Fixed: Variables not cleared in clearState()
+ * Show curtain when the section changes